refactor(): flour -> dough basin recipe creation
use a smarter method for adding these recipes by looping over the oreDict and then create the recipes this makes it easier if more salt gets added.

the flour oreDict is a convienence for the looping yet we only have one but this future proofs incase we change the flour mod.