Feat/ore changes (#98)
* refactor: wip changes to the ore processing and removals

note: still WIP and missing things

* chore: Reorganize docs

* chore: Move lingering text file and style with markdown

The file was just intended for planning, but I think it fits well as a documentation to reference. I've styled it with md so it looks a bit nicer too.

* refactor: re-add function call and remove old call

* fix: invalid conditional syntax

* style: Clean up commenting

* refactor: Remove zinc dust from manual staging

This is handled in the metals script

* style: Remove commented out item

This appears to be fine. It was used as a note as one point and it seems like the recipes its used in are fine..

* refactor: Move all ore oredict modifications to ores.zs

* refactor: Remove dustZinc oredict - now handled by metals

* fix: update gas method & furnace syntax

* refactor: add handling for speed & secondary outputs

necron will love this!

* fix: various recipe fixes & additions

* style: Add comments to better clarify items in array

* refactor: simplify adding secondary output to builder

* refactor: rename funcitons and variables in orematic

* fix: fix various script errors and complete recipe additions

also change the mk naming to match the other machines

* format: fix spacing on the orematic class

* fix: add temp recipe for dim ore

this will be removed/changed in 3.0.8

also add the start of mapping resources