refactor: Move unnamed recipes to seperate map
Although originally we were going to stick to an individual map, I think this results in cleaner code overall. This ends up reducing the amount of excess lines and work needed to add a recipe, and, unless its a unique situation, it will be very similar to the past implementation.