fix: Recipes to be staged later not staging
I did not see any recipe names staged unless they were created using the recipe util which generates a name that will match regex, so I believe this is all.

The recipe names will need to be staged after the IItemStack/IIngredient is recipe staged. Whichever method is done last is what the resulting recipes are staged in for that item.