fix(stages): tconstruct slime boots in multiple stages
As highlighted in #4158, the slime boots have inconsistencies the stages. Based on the way that the slime balls and blocks are staged, these were intended to be kept in stage 3, but the boots were also in stage 1.

The wildcard selector also causes a conflict and stages all of the boots to stage 1 as well, so this has been removed.

Closes #4158