bugfix(null-checks): Introduce sanity checks
In descriptions.zs and tooltips.zs:
You cannot set descriptions/tooltips on items that don't exist (e. g. environmentaltech:*). Is it a known issue?

In tinkers.zs and utils.zs:
When tinkers.zs/init passes a null object as `secondStage`, utils.zs/getHighestStage accesses `stage` on `secondStage`, resulting in a NullPointerException. I added an additional null check on stages and in case both are null, the function returns null as well.